Add test to check if translations are written correctly
